requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==9.5.0
psycopg2==2.9.6
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
WTForms==3.0.1
wtforms-validators==1.0.0
zipp==3.15.0
